Split riff and Istio charts
Because of how Helm manages CRDs, we cannot install Istio and riff at
the same time, they must be separate installs.

There are two Istio charts, istio and istio-lean. These are based on the
values used to generate the Istio installed in Knative. In short, istio
has a full mesh while istio-lean only supports ingress.

The Istio charts are the official Istio charts repackaged with custom
values and versioned with riff.

+++ package.sh
@@ -8,23 +8,34 @@ destination=$3
 
 mkdir -p charts/${chart}/templates
 
-while IFS= read -r line
-do
-  arr=($line)
-  name=${arr[0]%?}
-  url=${arr[1]}
-  args=$(echo $line | cut -d "#" -s -f 2)
-  file=charts/${chart}/templates/${name}.yml
+if [ -f charts/${chart}.yaml ] ; then
+  while IFS= read -r line
+  do
+    arr=($line)
+    name=${arr[0]%?}
+    url=${arr[1]}
+    args=$(echo $line | cut -d "#" -s -f 2)
+    file=charts/${chart}/templates/${name}.yml
 
-  curl -L -s ${url} > ${file}
+    curl -L -s ${url} > ${file}
 
-  # escape existing go template so helm doesn't get confused
-  cat ${file} | sed -e 's/{{/{{`{{/g' | sed -e 's/}}/}}`}}/g' > ${file}.tmp
-  mv ${file}.tmp ${file}
+    # escape existing go template so helm doesn't get confused
+    cat ${file} | sed -e 's/{{/{{`{{/g' | sed -e 's/}}/}}`}}/g' > ${file}.tmp
+    mv ${file}.tmp ${file}
 
-  # apply ytt overlays
-  ytt --ignore-unknown-comments -f overlays/ -f ${file} --file-mark $(basename ${file}):type=yaml-plain ${args} > ${file}.tmp
-  mv ${file}.tmp ${file}
-done < "charts/${chart}.yaml"
+    # apply ytt overlays
+    ytt --ignore-unknown-comments -f overlays/ -f ${file} --file-mark $(basename ${file}):type=yaml-plain ${args} > ${file}.tmp
+    mv ${file}.tmp ${file}
+  done < "charts/${chart}.yaml"
+fi
+
+if [ -f values/${chart}.yaml ] ; then
+  if [ -f charts/${chart}/values.yaml ] ; then
+    # merge custom values
+    yq merge -i -x charts/${chart}/values.yaml values/${chart}.yaml
+  else
+    cp values/${chart}.yaml charts/${chart}/values.yaml
+  fi
+fi
 
 helm package ./charts/${chart} --destination ${destination} --version ${version}
